REFACTOR: Split tree utils into 2 files
